requirements.txt
pandas==2.3.1
gspread==6.2.1
supabase==2.16.0
python_dotenv==1.1.1